+++ simple_gedcom/__init__.py
@@ -1,5 +1,5 @@
 from .parser import GedcomParser
-from .people import get_person_list, find_persons_by_name, save_person_list_to_csv 
+from .people import get_person_list, find_persons_by_name, save_person_list_to_csv
 from .pedigree import get_pedigree, save_pedigree_to_csv, show_generation_counts
 from .sources import get_source_list, get_person_source_list
 from .sources import save_source_list_to_csv, save_person_source_list_to_csv
@@ -7,7 +7,7 @@ from .gedcom_file import load_gedcom, GedcomFile
 from .marriages import get_marriages
 from .census import get_residence_list, get_census_list
 
-__version__ = "1.0.3"
+__version__ = "1.0.8"
 __all__ = ["GedcomParser", "GedcomFile", "load_gedcom",
            "get_person_list", "get_pedigree", "find_persons_by_name",
            "get_marriages",
@@ -15,5 +15,5 @@ __all__ = ["GedcomParser", "GedcomFile", "load_gedcom",
            "get_source_list", "get_person_source_list",
            "show_generation_counts",
            "save_person_list_to_csv", "save_pedigree_to_csv",
-           "save_source_list_to_csv", "save_person_source_list_to_csv"
+           "save_source_list_to_csv", "save_person_source_list_to_csv",
            ]